feat: enhance fake repositories and use cases with improved error handling and ID assignment

+++ apps/backend/src/modules/patients/application/usecases/DeletePatientUseCase.ts
@@ -10,8 +10,18 @@ export class DeletePatientUseCase {
   constructor(private repo: IPatientRepository) {}
 
   async execute(data: InputDTO) {
-    const patient = await this.repo.delete(data.id);
+    const exists = await this.repo.findById(data.id);
 
-    return patient;
+    if (!exists) {
+      throw new Error(`Patient not found`);
+      // throw new AppError(`Patient with ID ${data.id} not found`, 404);
+    }
+
+    try {
+      await this.repo.delete(data.id);
+    } catch (err) {
+      throw new Error('Could not delete procedure. Database error.');
+      // throw new AppError('Could not delete procedure. Database error.', 500);
+    }
   }
 }
