+++ apps/backend/src/modules/procedures/application/usecases/DeleteProcedureUseCase.ts
@@ -10,8 +10,18 @@ export class DeleteProcedureUseCase {
   constructor(private repo: IProcedureRepository) {}
 
   async execute(data: InputDTO) {
-    const procedure = await this.repo.delete(data.id);
+    const exists = await this.repo.findById(data.id);
 
-    return procedure;
+    if (!exists) {
+      throw new Error(`Procedure not found`);
+      // throw new AppError(`Procedure with ID ${data.id} not found`, 404);
+    }
+
+    try {
+      await this.repo.delete(data.id);
+    } catch (err) {
+      throw new Error('Could not delete procedure. Database error.');
+      // throw new AppError('Could not delete procedure. Database error.', 500);
+    }
   }
 }
